+++ WalletWasabi/Crypto/ZeroKnowledge/ZkVerifier.cs
@@ -15,18 +15,11 @@ namespace WalletWasabi.Crypto.ZeroKnowledge
 
 			var publicPoint = proof.PublicPoint;
 			var randomPoint = proof.RandomPoint;
-			var concatenation = ByteHelpers.Combine(
-				publicPoint.x.ToBytes(),
-				publicPoint.y.ToBytes(),
-				randomPoint.x.ToBytes(),
-				randomPoint.y.ToBytes());
-			using var sha256 = System.Security.Cryptography.SHA256.Create();
-			var hash = sha256.ComputeHash(concatenation);
-			var challenge = new Scalar(hash);
+			var challenge = ZkChallenge.Build(publicPoint, randomPoint);
 
-			var foo = (publicPoint * challenge + randomPoint).ToGroupElement();
-			var bar = (EC.G * proof.Response).ToGroupElement();
-			return (foo.IsInfinity && bar.IsInfinity) || (foo.x == bar.x && foo.y == bar.y);
+			var left = (publicPoint * challenge + randomPoint).ToGroupElement();
+			var right = (EC.G * proof.Response).ToGroupElement();
+			return (left.IsInfinity && right.IsInfinity) || (left.x == right.x && left.y == right.y);
 		}
 	}
 }
